requirements.txt
Flask==2.3.2
requests>=2.32.0
pytest==7.2.2
python-dotenv==1.0.0